Source files rename for cli classes.
