+++ src/smack_cli.hpp
