+++ test/cli_class.cpp
@@ -11,7 +11,7 @@
 
 #include "test_common.hpp"
 
-#include "../src/smack_util_console_app.hpp"
+#include "../src/smack_cli.hpp"
 
 using std::cout;
 using std::endl;
